+++ src/collector/rrdp.rs
@@ -16,7 +16,8 @@
 use std::{cmp, error, fmt, fs, io, mem};
 use std::collections::{HashSet, HashMap};
 use std::convert::{TryFrom, TryInto};
-use std::path::Path;
+use std::io::Write;
+use std::path::{Path, PathBuf};
 use std::sync::{Arc, Mutex, RwLock};
 use std::time::{Duration, SystemTime};
 use bytes::Bytes;
@@ -320,7 +321,7 @@ impl<'a> Run<'a> {
             Some(ref http) => http,
             None => return None,
         };
-        let mut response = match http.response(uri) {
+        let mut response = match http.response(uri, false) {
             Ok(response) => response,
             Err(_) => return None,
         };
@@ -605,7 +606,7 @@ impl Repository {
                 }
             };
             Some((uri, object.content))
-        }).filter_map(|item| item)
+        }).flatten()
     }
 }
 
@@ -712,7 +713,9 @@ impl<'a> RepositoryUpdate<'a> {
     ) -> Result<(), SnapshotError> {
         debug!("RRDP {}: updating from snapshot.", self.rpki_notify);
 
-        let response = match self.http.response(notify.snapshot.uri()) {
+        let response = match self.http.response(
+            notify.snapshot.uri(), false
+        ) {
             Ok(response) => {
                 self.metrics.payload_status = Some(response.status().into());
                 response
@@ -910,7 +913,7 @@ impl<'a> RepositoryUpdate<'a> {
         uri: &uri::Https,
         hash: rrdp::Hash,
     ) -> Result<sled::Batch, DeltaError> {
-        let response = match self.http.response(uri) {
+        let response = match self.http.response(uri, false) {
             Ok(response) => {
                 self.metrics.payload_status = Some(response.status().into());
                 response
@@ -955,6 +958,9 @@ struct HttpClient {
     /// This will be of the error variant until `ignite` has been called. Yes,
     /// that is not ideal but 
     client: Result<Client, Option<ClientBuilder>>,
+
+    /// The base directory for storing copies of responses if that is enabled.
+    response_dir: Option<PathBuf>,
 }
 
 impl HttpClient {
@@ -1010,6 +1016,7 @@ impl HttpClient {
         }
         Ok(HttpClient {
             client: Err(Some(builder)),
+            response_dir: config.rrdp_keep_responses.clone(),
         })
     }
 
@@ -1079,11 +1086,19 @@ impl HttpClient {
     }
 
     /// Performs an HTTP GET request for the given URI.
+    ///
+    /// If keeping responses is enabled, the response is written to a file
+    /// corresponding to the URI. If the resource behind the URI changes over
+    /// time and this change should be tracked, set `multi` to `true` to
+    /// include the current time in the file name.
     pub fn response(
         &self,
-        uri: &uri::Https
-    ) -> Result<Response, reqwest::Error> {
-        self.client().get(uri.as_str()).send()
+        uri: &uri::Https,
+        multi: bool,
+    ) -> Result<HttpResponse, reqwest::Error> {
+        self.client().get(uri.as_str()).send().map(|response| {
+            HttpResponse::create(response, uri, &self.response_dir, multi)
+        })
     }
 
     /// Requests, parses, and returns the given RRDP notification file.
@@ -1097,7 +1112,7 @@ impl HttpClient {
         uri: &uri::Https,
         status: &mut HttpStatus,
     ) -> Option<NotificationFile> {
-        let response = match self.response(uri) {
+        let response = match self.response(uri, true) {
             Ok(response) => {
                 *status = response.status().into();
                 response
@@ -1125,6 +1140,118 @@ impl HttpClient {
                 None
             }
         }
+    }
+}
+
+
+//------------ HttpResponse --------------------------------------------------
+
+/// Wraps a reqwest response for added features.
+struct HttpResponse {
+    /// The wrapped reqwest response.
+    response: Response,
+
+    /// A file to also store read data into.
+    file: Option<fs::File>,
+}
+
+impl HttpResponse {
+    /// Creates a new response wrapping a reqwest reponse.
+    ///
+    /// If `response_dir` is some path, the response will also be written to
+    /// a file under this directory based on `uri`. Each URI component
+    /// starting with the authority will be a directory name. If `multi` is
+    /// `false` the last component will be the file name. If `multi` is
+    /// `true` the last component will be a directory, too, and the file name
+    /// will be the ISO timestamp of the current time.
+    pub fn create(
+        response: Response,
+        uri: &uri::Https,
+        response_dir: &Option<PathBuf>,
+        multi: bool
+    ) -> Self {
+        HttpResponse {
+            response,
+            file: response_dir.as_ref().and_then(|base| {
+                Self::open_file(base, uri, multi).ok()
+            })
+        }
+    }
+
+    /// Opens the file mirroring file.
+    ///
+    /// See [`create`][Self::create] for the rules.
+    fn open_file(
+        base: &Path, uri: &uri::Https, multi: bool
+    ) -> Result<fs::File, Failed> {
+        let path = base.join(&uri.as_str()[8..]);
+        let path = if multi {
+            path.join(Utc::now().to_rfc3339())
+        }
+        else {
+            path
+        };
+
+        let parent = match path.parent() {
+            Some(parent) => parent,
+            None => {
+                warn!(
+                    "Cannot keep HTTP response; \
+                    URI translated into a bad path '{}'",
+                    path.display()
+                );
+                return Err(Failed)
+            }
+        };
+        if let Err(err) = fs::create_dir_all(&parent) {
+            warn!(
+                "Cannot keep HTTP response; \
+                creating director {} failed: {}",
+                parent.display(), err
+            );
+            return Err(Failed)
+        }
+        fs::File::create(&path).map_err(|err| {
+            warn!(
+                "Cannot keep HTTP response; \
+                creating file {} failed: {}",
+                path.display(), err
+            );
+            Failed
+        })
+    }
+
+    /// Returns the value of the content length header if present.
+    pub fn content_length(&self) -> Option<u64> {
+        self.response.content_length()
+    }
+
+    /// Copies the full content of the response to the given writer.
+    pub fn copy_to<W: io::Write + ?Sized>(
+        &mut self, w: &mut W
+    ) -> Result<u64, io::Error> {
+        // We cannot use the reqwest response’s `copy_to` impl because we need
+        // to use our own `io::Read` impl which sneaks in the copying to file
+        // if necessary.
+        io::copy(self, w)
+    }
+
+    /// Returns the status code of the response.
+    pub fn status(&self) -> StatusCode {
+        self.response.status()
+    }
+}
+
+
+//--- Read
+
+impl io::Read for HttpResponse {
+    fn read(&mut self, buf: &mut [u8]) -> Result<usize, io::Error> {
+        let res = self.response.read(buf)?;
+        if let Some(file) = self.file.as_mut() {
+            file.write_all(&buf[..res])?;
+        }
+        Ok(res)
     }
 }