+++ Changelog.md
@@ -16,6 +16,10 @@ Breaking Changes
   now fall back to rsync. The time since last successful update before
   this fallback happens is configurable via the `rrdp-fallback-time`
   option and defaults to one hour. ([#473], [#482])
+* Routinator now includes additional TALs for various commonly used
+  RPKI testbeds. The `init` command has been restructured to make it
+  possible to select the TALs for installation. The default is still to
+  install the five production RIR TALs. ([#500])
 * Deprecated configuration items have been removed: `unknown-objects`
   cannot be spelled with a underscore anymore and string values are not
   accepted anymore for `rtr-tcp-keepalive`. ([#496])
@@ -37,6 +41,8 @@ New
   ([#463], contributed by [@bjpbakker])
 * The `exception` config file value now also accepts a single string with
   a path name instead of an array of strings. ([#471])
+* The new `rrdp-keep-responses` option allows optionally storing the XML
+  content of all received RRDP responses in the file system. ([#490])
 * The HTTP endpoints that supply the current VRP set now support
   conditional request handling. They include Etag and
   Last-Modified headers in their response and process If-None-Match and
@@ -72,8 +78,10 @@ Other Changes
 [#487]: https://github.com/NLnetLabs/routinator/pull/487
 [#488]: https://github.com/NLnetLabs/routinator/pull/488
 [#489]: https://github.com/NLnetLabs/routinator/pull/489
+[#490]: https://github.com/NLnetLabs/routinator/pull/490
 [#496]: https://github.com/NLnetLabs/routinator/pull/496
 [#498]: https://github.com/NLnetLabs/routinator/pull/498
+[#500]: https://github.com/NLnetLabs/routinator/pull/500
 [#505]: https://github.com/NLnetLabs/routinator/pull/505
 [rpki-rs]: https://github.com/NLnetLabs/rpki-rs/
 [rpki-rtr]: https://github.com/NLnetLabs/rpki-rtr/